Resolve build errors

+++ src/melee/lb/lbarchive.h
@@ -6,7 +6,7 @@
 
 #include <baselib/archive.h>
 
-void lbArchive_InitializeDAT(HSD_Archive* archive, u8* data, u32 length);
+void lbArchive_InitializeDAT(HSD_Archive* archive, u32* data, u32 length);
 void lbArchive_80016EFC(void);
 HSD_Archive* lbArchive_LoadArchive(char* filename);
